test(*): write initial pass of tests
These tests use the starlette TestClient to mock the server being exposed. There is also the
assumption that the database is not mocked and is created and destroyed for each test suite. This
database is hosted on the default server set through environment variables and expects the database
to be called 'test'. Therefore it is required that the following environment variable be exported
before running tests: POSTGRES_DB=test.