fix(model schema): improve model schema
I made a number of changes to model schema to make it more restrict. Here is the list:
1.
rad_modifier shouldn't be required and will be set based on Face.face_type. I tried to implement the
default value based on face_type and it is totally possible but mixes up the openapi documentation
so I took it out. We can take care of this by running model through a default value generator as
discussed in https://github.com/pollination/model-service/issues/8
2. apertures do not need parent
object. We already know the parent which is the Face.
3. Face needs a parent which will be a zone
and is optional and only needed for energy modeling.
4. Added a validator for vertices length. Also
changed `min_items` to `minItems`.
5. Removed `Window` from FaceType enumeration and renamed it to
FaceTypeOpaque.
6. Limited aperture rad_modifier to Glass and Transparent.

BREAKING CHANGE: This change will break old schema which has a parent field for apertures.