refactor(tests): Refactor tests to all use JSON files
The Python dictionaries were following an alternate method of testing that could easily get out of sync with the JSON samples that we are pointing people towards as a reference for the schema.  This commit gets rid of these Python files and has all of the tests use the JSON files.